refactor(cli): share immutable Runtime Host package store
